Split out storage-related types.
