+++ worker/uniter/runner/jujuc/testing/storage.go
@@ -86,40 +86,3 @@ func (c *ContextStorage) AddUnitStorage(all map[string]params.StorageConstraints
 
 	c.info.AddUnitStorage(all)
 }
-
-// StorageAttachment holds the data for the test double.
-type StorageAttachment struct {
-	Tag      names.StorageTag
-	Kind     storage.StorageKind
-	Location string
-}
-
-// ContextStorageAttachment is a test double for jujuc.ContextStorageAttachment.
-type ContextStorageAttachment struct {
-	contextBase
-	info *StorageAttachment
-}
-
-// Tag implements jujuc.StorageAttachement.
-func (c *ContextStorageAttachment) Tag() names.StorageTag {
-	c.stub.AddCall("Tag")
-	c.stub.NextErr()
-
-	return c.info.Tag
-}
-
-// Kind implements jujuc.StorageAttachement.
-func (c *ContextStorageAttachment) Kind() storage.StorageKind {
-	c.stub.AddCall("Kind")
-	c.stub.NextErr()
-
-	return c.info.Kind
-}
-
-// Location implements jujuc.StorageAttachement.
-func (c *ContextStorageAttachment) Location() string {
-	c.stub.AddCall("Location")
-	c.stub.NextErr()
-
-	return c.info.Location
-}
